% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReigenMT.R
\name{eigenmt}
\alias{eigenmt}
\title{EigenMT: Multiple Testing Correction for Summary Statistics}
\usage{
eigenmt(
  summary_stats,
  genotypes,
  genotype_to_position = NULL,
  variant_column_summary_stats = "SNP",
  feature_column_summary_stats = "gene",
  pvalue_column = "p-value",
  genotype_position_column = "pos",
  genotype_var_position_column = "snp",
  var_explained_threshold = 0.975,
  window_size = 100
)
}
\arguments{
\item{summary_stats}{A data.table or data.frame containing summary statistics.}

\item{genotypes}{A data.table, data.frame, matrix, or list containing genotype data.}

\item{genotype_to_position}{A data.table, data.frame, list, or hash table mapping genotypes to positions. If not supplied, and the genotype object is plink, it can be created on-the-fly. Default is NULL.}

\item{variant_column_summary_stats}{A string specifying the column name for variants in the summary statistics. Default is 'SNP'.}

\item{feature_column_summary_stats}{A string specifying the column name for features in the summary statistics. Default is 'gene'.}

\item{pvalue_column}{A string specifying the column name for p-values in the summary statistics. Default is 'p-value'.}

\item{genotype_position_column}{A string specifying the column name for positions in the genotype data. Default is 'pos'.}

\item{genotype_var_position_column}{A string specifying the column name for variant positions in the genotype data. Default is 'snp'.}

\item{var_explained_threshold}{A numeric value indicating the threshold of variance to be explained. Default is 0.975.}

\item{window_size}{An integer specifying the window size for checking multiple variants. Default is 100.}
}
\value{
A data.table with the corrected summary statistics, including the number of tests per feature, the bonferroni corrected p-value considering the number of tests of that feature, and the bonferroni corrected p-value considering all independent tests across all features
}
\description{
This function performs multiple testing correction on summary statistics using eigenvalues of genotype data.
}
\examples{
\dontrun{
summary_stats <- data.table(SNP = c("rs1", "rs2"), gene = c("gene1", "gene2"), `p-value` = c(0.01, 0.05))
genotypes <- data.table(ID = c("var1", "var2"), value = c(1, 2))
genotype_to_position <- data.table(snp = c("rs1", "rs2"), pos = c(100, 200))
corrected_stats <- eigenmt(summary_stats, genotypes, genotype_to_position)
print(corrected_stats)
}
}
