% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReigenMT.R
\name{find_number_of_eigen}
\alias{find_number_of_eigen}
\title{Find Number of Eigenvalues to Explain Variance}
\usage{
find_number_of_eigen(
  eigenvalues,
  n_variants,
  var_explained_threshold = 0.975,
  verbose = F
)
}
\arguments{
\item{eigenvalues}{A numeric vector of eigenvalues.}

\item{n_variants}{An integer specifying the number of variants.}

\item{var_explained_threshold}{A numeric value indicating the threshold of variance to be explained. Default is 0.975.}

\item{verbose}{A logical value indicating whether to print detailed messages. Default is FALSE.}
}
\value{
An integer representing the number of eigenvalues needed to explain the specified variance threshold.
}
\description{
This function determines the number of eigenvalues needed to explain a specified threshold of variance.
}
\examples{
\dontrun{
eigenvalues <- c(2.5, 1.5, 1.0, 0.5)
n_variants <- 4
threshold <- 0.975
num_eigen <- find_number_of_eigen(eigenvalues, n_variants, threshold)
print(num_eigen)
}
}
