% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReigenMT.R
\name{genotypes_to_eigenvalues}
\alias{genotypes_to_eigenvalues}
\title{Compute Eigenvalues from Genotype Data}
\usage{
genotypes_to_eigenvalues(genotypes, verbose = F)
}
\arguments{
\item{genotypes}{A data.table, data.frame, matrix, or list containing genotype data.}

\item{verbose}{A logical value indicating whether to print detailed messages. Default is FALSE.}
}
\value{
A vector of eigenvalues of the shrinkage covariance matrix.
}
\description{
This function computes the eigenvalues of the regularized (shrinkage) covariance matrix from genotype data.
}
\examples{
\dontrun{
genotypes <- data.table(ID = c("var1", "var2", "var3"), value = c(1, 2, 3))
eigenvalues <- genotypes_to_eigenvalues(genotypes)
print(eigenvalues)
}
}
