% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReigenMT.R
\name{get_position_hashtab}
\alias{get_position_hashtab}
\title{Create a Position Hash Table}
\usage{
get_position_hashtab(feature_to_position, var_column, pos_column)
}
\arguments{
\item{feature_to_position}{A data.table, data.frame, list, or other recognized format containing feature and position information.}

\item{var_column}{The name of the column containing feature names.}

\item{pos_column}{The name of the column containing position values.}
}
\value{
A hash table mapping each feature to its corresponding position.
}
\description{
This function creates a hash table mapping features to their positions from various input formats. This table is required in some functions, to link variants to genomic positions.
}
\examples{
\dontrun{
feature_data <- data.table(feature = c("feature1", "feature2"), position = c(100, 200))
hash_table <- get_position_hashtab(feature_data, "feature", "position")
print(hash_table)
}
}
