% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReigenMT.R
\name{limix_h5_to_sumstats_format}
\alias{limix_h5_to_sumstats_format}
\title{Convert HDF5 to Summary Statistics Format}
\usage{
limix_h5_to_sumstats_format(h5_loc)
}
\arguments{
\item{h5_loc}{A string specifying the location of the HDF5 file.}
}
\value{
A data.table containing the merged data from all features in the HDF5 file,
with an additional column named 'feature' indicating the feature name.
}
\description{
This function reads an HDF5 file and converts its contents into a data.table
with an additional column indicating the feature name. This assumes the format of chunked LIMIX-QTL outputs
}
\examples{
\dontrun{
h5_file <- "path/to/your/file.h5"
result <- limix_h5_to_sumstats_format(h5_file)
print(result)
}
}
