% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReigenMT.R
\name{subset_genotypes}
\alias{subset_genotypes}
\title{Subset Genotypes by Variants}
\usage{
subset_genotypes(genotypes, variants, id_column = "ID", strip_id = T)
}
\arguments{
\item{genotypes}{A data.table, data.frame, matrix, or list containing genotype data.}

\item{variants}{A vector of variant IDs to subset the genotypes.}

\item{id_column}{A string specifying the name of the column containing variant IDs. Default is 'ID'.}

\item{strip_id}{A boolean whether or not to remove the identifier from the table. Default is True.}
}
\value{
A subset of the genotype data containing only the specified variants, with the ID column removed.
}
\description{
This function subsets genotype data based on a list of variants and optionally removes the ID column.
}
\examples{
\dontrun{
genotypes <- data.table(ID = c("var1", "var2", "var3"), value = c(1, 2, 3))
variants <- c("var1", "var3")
subset <- subset_genotypes(genotypes, variants, "ID")
print(subset)
}
}
